restructure testfiles
